Switch default binary to bootroot
Closes #59
